Speed up "Install package" task
Closes #155

CHANGES:

Before this patch, instances for the "Install Package" task were selected on each instance, it was too long.

Now the list of instances for installing a package is selected once for all.
This made it possible to speed up the task at times for a large cluster.

COMPLEXITY AND TIME:

Before this patch, the complexity of calculating the list of instances was `O(N^2)`, now it is `O(N)`.
For 100 instances, this gives a 10x time reduction, from 60 seconds to 5 seconds.