Split role into sections (#178)
Closes #141
Closes #135

Before this patch, the `tarantool.cartridge` role used tags to highlight some
of the global stages. This allowed it to be partially used by embedding it
in other processes. But it was impossible to change, for example, the stage of
loading an artifact.

It is proposed to divide the role into many steps and combine them using
a scenario in the config. Accordingly, this scenario can be rewritten via `vars`,
which allows you to run a role several times in one playbook with different scenarios.

To use custom tasks in the role configuration, it is proposed to specify the folder
with user steps through the option in the configuration. YAML files will be
searched in the specified folder, and then the names of the files will be taken
for the step names: the file `./tasks/my_task.yml` will be imported as the step
`my_task`. These steps can also be used in `cartridge_scenario`.

It is also proposed to add the ability to import files with custom tasks from arbitrary
places with arbitrary names. These steps can also be used in `cartridge_scenario`.

If the step names conflict, the step with a higher priority will be executed
(allows you to overwrite any role step with a custom one).
Priority:
- the step described in `cartridge_custom_steps`;
- step from the folder specified in `cartridge_custom_steps_dir`;
- role step.