Fix and improve `needs_restart`
Closes #189

In the current realization `cartridge_needs_restart` is placed in `configure_instance` step.
It checks that instance code is newer that socket (package was updated after instance start)
AND
that instance configuration changes require instance restart to be applied.

This approach is wrong: in case of [update_instance, restart_instance] scenario
we skip this check and instance wouldn't be restarted.

Now `needs_restart` variable is recalculated on each step that can require instance restart:
- `update_package`
- `update_instance`
- `configure_instance`