Add some default scenarios and availability to define custom scenarios
Closes #217

Before scenarios were added, tags were used to partially run of role.
Now, with the scenarios added, you can decide for yourself which steps you want to start.
But sometimes new steps can be added to the default script.
If a step is added to an existing logical group of steps
(for example, a new step for configuring an application),
then those users who use their scenarios (for example, for configuring an application)
will not have this step and they will have to add this step manually.

Let's make some default scripts (for example, "package update", "topology configuration", "application configuration").
And the user will be able to choose one of the supported scenarios.