Add script to count test matrix
Before this patch, there was GitHub Actions code duplication
as it was difficult to combine the same test with different parameters.

Now a script has been added that generates the test matrix,
so duplication of the code has been removed.

Also, most of the tests were turned off by default.
To start them, you should call the event "workflow_dispatch".