Fix facts setting in `hostvars` fact (#278)
Before this patch, complex variables (lists, dictionaries) were not fully expanded
on old Ansible versions (2.8 or earlier). As a result, it was impossible to use other
variables (e.g. `{{ common_memtx }}` value) in dictionaries and lists in `hosts.yml`.

We are now using the standard `set_fact` mechanism, which solves this problem.