Speed up tasks that use `hostsvars`
Closes #260

Beforw this patch the `hostsvars` variable is too large when deploying huge clusters.
Therefore, tasks that use `hostsvars` are executed for at least 40 seconds when deploying 500 instances.

Now only the necessary information will be transferred.
Due to this we will reduce duration of the tasks which use `hostvars`.

Some duration reduction results:
- Select control instance to manage topology and configuration from `70.38s` to `22.85s`
- Connect instance to membership from `54.30s` to `1.28s`
- Select one not expelled instance from `37.46s` to `0.71s`
- Select one instance for each physical machine from `30.60s` to `0.43s`
- Validate config from `26.03s` to `2.02s`