Consider enabled roles on editing topology
Before this patch passed `roles` was compared with `replicaset.roles`.
But `replicaset.roles` contain roles and their dependencies, so it can
be different from passed `roles`. This leads to unneseccary
`edit_topology` calls for all replicasets, that leads to different
isuues on big clusters. Since this patch enabled roles for passed
`roles` and `replicaset.roles` are compared to understand if changes
are needed.