Consider enabled roles on editing topology (#318)
Before this patch passed `roles` were compared with `replicaset.roles`.
But `replicaset.roles` contain roles and their dependencies, so it can
be different from passed `roles`. This leads to unnecessary
`edit_topology` calls for all replicasets, which leads to different
issues on big clusters. Since this patch enabled roles for passed
`roles` and `replicaset.roles` are compared to understand if changes
are needed.