Make all retries and delays configurable
Closes #186

In some tasks we have to specify number of retries and delays between to wait for some operation to finish.
Seems that this constants shound be moved to defaults and also these constants should be overridable by extra-vars.

Also in some tasks timeouts are used instead retries and delays.

Now all retries and delays are configurable and some timeouts deprecated.