Add step to compare inventory and real cluster topology
Closes #215
Closes #362

Before the patch, it was dangerous to edit topology.

Now, in `edit_topology` step has been added the check, which compares the current
topology from the working cluster with inventory and if all dangerous changes check
passes only then it will continue to play.
It's possible to call this check by the `check_edit_topology` step.
Also, it's possible to disable checks by flags.