Optimize 'Cache facts' step
The fact caching step was actually filtering the hostvars dictionary.

Before this patch, the Ansible module was used for this step.
As a result, Ansible spent 90% of execution time to create an environment
for calling the module (for 500 instances the call takes 20-30 seconds).

Now the Jinja2 filter is used to create a dictionary of cached facts,
which has accelerated the call of the filtering function by 10 times
(for 500 instances the call takes 2-4 seconds).