Use `os.lstat` instead of `os.path.getmtime`
Closes #383

Sometimes a situation arises that it is necessary to roll back an instance
to an old version of the TGZ package. After updating the package link,
the instance should be rebooted.

Before this patch, the `os.path.getmtime` function was used, which did
not check the modification time of the link, but the modification time of the package,
so the instance did not reboot (since the package is older than the restart time).

Now the `os.lstat` function is used, which does not follow symbolic links and
looks at the attributes of the link itself.