Remove old configs before loading a new one
Closes #406

Before the patch, the new config was loaded into the `/tmp` folder and stayed there.
As a result, when loading the folder config with the old name,
there were files from the old config that were deleted in the new config.

Now all configs are loaded into the `/tmp/tarantool.cartridge.configs/` folder,
which allows to remove previous configs before loading a new one.